Final refactorization
Delete TemplateObjectAr class.
